Update go-telegram/bot version to v1.1.7 and add AllowedChatIDsList

+++ sender/telegram.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"log/slog"
+	"slices"
 	"sync"
 	"time"
 
@@ -33,6 +34,27 @@ func InitSender(lgr *slog.Logger, config *conf.Config) (*Sender, error) {
 	opts := []bot.Option{
 		bot.WithDefaultHandler(sender.handler),
 		bot.WithSkipGetMe(),
+		// list of alloweed updates
+		// https://core.telegram.org/bots/api#update
+		bot.WithAllowedUpdates(bot.AllowedUpdates{
+			// "callback_query",
+			// "channel_post",
+			// "chat_boost",
+			"chat_join_request",
+			"chat_member",
+			// "chat_migration",
+			// "chosen_inline_result",
+			// "edited_channel_post",
+			// "edited_message",
+			// "inline_query",
+			// "message",
+			// "message_reaction",
+			// "message_reaction_count",
+			"my_chat_member",
+			// "poll",
+			// "poll_answer",
+			// "removed_chat_boost",
+		}),
 	}
 
 	// if config.Debug {
@@ -54,13 +76,17 @@ func InitSender(lgr *slog.Logger, config *conf.Config) (*Sender, error) {
 
 func (s *Sender) handler(ctx context.Context, b *bot.Bot, update *bm.Update) {
 	if s.config.Debug {
-		if update.Message != nil && update.Message.From != nil && update.Message.Chat.ID != 0 && update.Message.Text != "" {
-			s.lgr.Debug(fmt.Sprintf("%d -> %d: %s", update.Message.From.ID, update.Message.Chat.ID, update.Message.Text))
-		}
+		s.lgr.Debug(formatUpdateForLog(update))
 	}
 
 	if s.config.RestictOnJoin && update.Message != nil && update.Message.NewChatMembers != nil {
-		s.lgr.Debug(fmt.Sprintf("restrict users %#v", update.Message.NewChatMembers))
+		s.lgr.Info(fmt.Sprintf("Restrict users %#v", update.Message.NewChatMembers))
+
+		if !slices.Contains(s.config.AllowedChatIDsList, update.Message.Chat.ID) {
+			s.lgr.Info(fmt.Sprintf("Chat ID %d is not in allowed list", update.Message.Chat.ID))
+			return
+		}
+
 		for _, member := range update.Message.NewChatMembers {
 			_, err := s.Bot.RestrictChatMember(
 				context.Background(),
@@ -82,6 +108,11 @@ func (s *Sender) handler(ctx context.Context, b *bot.Bot, update *bm.Update) {
 	if s.config.DeleteJoinMessages && update.Message != nil && update.Message.NewChatMembers != nil {
 		s.lgr.Info(fmt.Sprintf("Member joined %+v, chat ID %d", update.Message.NewChatMembers, update.Message.Chat.ID))
 
+		if !slices.Contains(s.config.AllowedChatIDsList, update.Message.Chat.ID) {
+			s.lgr.Info(fmt.Sprintf("Chat ID %d is not in allowed list", update.Message.Chat.ID))
+			return
+		}
+
 		_, err := s.Bot.DeleteMessage(
 			context.Background(),
 			&bot.DeleteMessageParams{
@@ -98,7 +129,12 @@ func (s *Sender) handler(ctx context.Context, b *bot.Bot, update *bm.Update) {
 	}
 
 	if s.config.DeleteLeaveMessages && update.Message != nil && update.Message.LeftChatMember != nil {
-		s.lgr.Info(fmt.Sprintf("Member left %+v, chat ID %d", update.Message.LeftChatMember, update.Message.Chat.ID))
+		s.lgr.Info(fmt.Sprintf("Member has left %+v, chat ID %d", update.Message.LeftChatMember, update.Message.Chat.ID))
+
+		if !slices.Contains(s.config.AllowedChatIDsList, update.Message.Chat.ID) {
+			s.lgr.Info(fmt.Sprintf("Chat ID %d is not in allowed list", update.Message.Chat.ID))
+			return
+		}
 
 		_, err := s.Bot.DeleteMessage(
 			context.Background(),
@@ -114,6 +150,4 @@ func (s *Sender) handler(ctx context.Context, b *bot.Bot, update *bm.Update) {
 
 		return
 	}
-
-	s.lgr.Debug(fmt.Sprintf("Message %#v", update))
 }
